feat: Add obj-input-variable for form-component.
This commit adds a new input-variable next to the form_data
input_variable. The new input_variable is named after the selected model
and uses the lowercase / snakecase name. It will contain the record_id
as value if present, with the key `variable_id`.
B5 uses this variable_id to fetch the correct record.

Example:

Lets say we use the model Webuser. Then the input_variables will look
like this:

{form_data: {construct_id: abc, id: 123}, webuser{variable_id: 123}}