Fix(PAGE-2170): Remove afterDelete 'deleteActionVariable' from components.
Components that have been duplicated, by duplicate-page for example,
point to the original ActionVariable, instead of the duplicate
ActionVariable. Meaning that after delete, a mutation to the MetaAPI is
been made to delete the wrong actionVariable.

This commit removes the afterDelete deleteActionVariable from the
components.

In the MetaAPI, we have applied changes - with the same jira-tag:
PAGE-2170 - which changed the logic of deleting an component. Now, when
you delete a component, and the component points to an actionVariable,
it will also delete the actionVariable. The actionVariableId used in the
MetaAPI points to the correct actionVariable: the duplicated one.