fix: remove all instances of betty-jwt cookie
After logging in via username/password, or custom-login, the betty-jwt
cookie needs to always be deleted.
If the cookie was set via a page with an extra path in its url, the
login-function wouldn't delete it, and the user would need to login
twice.
By explicitly setting the path when deleting the cookie, the cookie is
always deleted thus solving the bug.