fix: rely on use-page-state to get the current value
The reversed code solved a bug that only occured when the
input-component was iterated. When a component is iterated, it would
trigger some code in the compiler that didn't perfectly handle the
state.

But the solve broke a use-case, where component is in a conditional, and
the state is persisted when the component is temporarily not on the
page.

By reverting this code, and fixing the compiler, all use-cases now work.